# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Influencer Travel Marketplace — AI-Powered Itinerary Planning API
  description: |
    ## Overview

    REST API for the **Influencer Travel Marketplace** platform.
    This API enables three user roles to interact with the marketplace:

    | Role | Capabilities |
    |------|-------------|
    | `influencer` | Create and manage travel packages, upload media, use AI itinerary suggestions |
    | `admin` | Review, approve, and reject packages submitted for marketplace listing |
    | `customer` | Browse the marketplace, search for packages, view package details |

    ## Architecture

    - **Backend:** FastAPI (Python), deployed on Railway *— Phase 2 target; Phase 1 is Flask 3.0.3 on port 5000 with no `/v1` prefix. See [apps/api/docs/adr/0001-flask-now-fastapi-for-ai.md](../apps/api/docs/adr/0001-flask-now-fastapi-for-ai.md).*
    - **Database:** Supabase PostgreSQL (13 tables)
    - **AI:** LangChain RAG pipeline with GPT-4o (NFR-P-101: responses < 10 seconds)
    - **Frontend:** Next.js + Tailwind CSS, deployed on Vercel
    - **Authentication:** JWT Bearer tokens issued on login

    ## Package Lifecycle

    ```
    draft → pending_review → approved → live → archived
                          ↘ rejected → draft (via revise)
    ```

    ## Non-Functional Requirements

    - **NFR-P-100:** All user-facing text is localised to English (Australia); prices in whole AUD dollars.
...

    ## Assessor Notes

...

  version: "2.0.0"
  license:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /auth/login`. Include in every protected request:
        ```
        Authorization: Bearer <token>
        ```
        The token payload contains:
        ```json
        {
          "sub": "uuid",
          "role": "influencer | admin | customer",
          "email": "user@example.com",
          "exp": 1748822400
        }
        ```

  # ─────────────────────────────────────────────
  #  REUSABLE SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    TravelPackageSummary:
      type: object
      properties:
        package_id:
          type: string
          format: uuid
          example: "pkg-00001-uuid-example-here"
        title:
          type: string
          example: "10-Day Bali Escape: Culture, Surf & Cuisine"
        destination_country:
          type: string
          nullable: true
          example: "Indonesia"
        destination_city:
          type: string
          nullable: true
          example: "Bali"
        duration_days:
          type: integer
          nullable: true
          example: 10
        base_price_aud:
          type: integer
          example: 3500
        status:
          $ref: '#/components/schemas/PackageStatus'
        creator_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        submitted_at:
          type: string
          format: date-time
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        cover_image_url:
          type: string
          format: uri
          nullable: true
          example: "https://storage.supabase.co/object/public/media/pkg-00001/cover.jpg"
    TravelPackageDetail:
      allOf:
        - $ref: '#/components/schemas/TravelPackageSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            max_group_size:
              type: integer
              nullable: true
            tags:
              type: array
              items:
                type: string
            flights:
              type: array
              items:
                $ref: '#/components/schemas/FlightDetail'
            hotels:
              type: array
              items:
                $ref: '#/components/schemas/HotelDetail'
            activities:
              type: array
              items:
                $ref: '#/components/schemas/ActivityDetail'
            media:
              type: array
              items:
                $ref: '#/components/schemas/MediaItem'
            days:
              type: array
              description: Per-day narrative for detail-page rendering, ordered by day_number. Group flights/hotels/activities client-side via their day_number / check_in_day fields.
              items:
                $ref: '#/components/schemas/PackageDay'
            creator:
              $ref: '#/components/schemas/PackageCreator'
              nullable: true
            latest_approval:
              $ref: '#/components/schemas/ApprovalRecord'
              nullable: true
            pricing:
              $ref: '#/components/schemas/Pricing'
              nullable: true
    Pricing:
      type: object
      description: |
        Informational price breakdown computed from the package's itinerary
        components (null component prices and null hotel nights count as 0).
        `base_price_aud` is the
        creator-set display price and may differ from `components_total`; the
        API asserts nothing about which is authoritative.
      properties:
        flights_total:
          type: integer
          description: Sum of flight prices in whole AUD.
        hotels_total:
          type: integer
          description: Sum of price-per-night × nights across hotels, whole AUD.
        activities_total:
          type: integer
          description: Sum of activity prices in whole AUD.
        components_total:
          type: integer
          description: flights_total + hotels_total + activities_total.
        base_price_aud:
          type: integer
          nullable: true
          description: The package's creator-set display price, repeated for convenience.
    PackageStatus:
      type: string
      enum: [draft, pending_review, approved, rejected, live, archived]
      description: |
        Package lifecycle states:
        - `draft` — being edited by influencer
        - `pending_review` — submitted, awaiting admin decision
        - `approved` — admin approved, ready to publish
        - `rejected` — admin rejected; influencer can revise and resubmit
        - `live` — published and visible in marketplace
        - `archived` — removed from marketplace
...
    FlightInput:
      type: object
      required: [origin_iata, destination_iata, airline, departure_datetime, arrival_datetime]
      properties:
        origin_iata:
          type: string
          minLength: 3
          maxLength: 3
          example: "SYD"
        destination_iata:
          type: string
          minLength: 3
          maxLength: 3
          example: "DPS"
        airline:
          type: string
          example: "Garuda Indonesia"
        flight_number:
          type: string
          nullable: true
          example: "GA715"
        departure_datetime:
          type: string
          format: date-time
          example: "2025-07-14T06:30:00+10:00"
        arrival_datetime:
          type: string
          format: date-time
          example: "2025-07-14T11:45:00+08:00"
        cabin_class:
          type: string
          enum: [economy, premium_economy, business, first]
          example: "economy"
        price_aud:
          type: integer
          description: Per-person flight price in whole AUD dollars.
          example: 850
    FlightDetail:
      allOf:
        - type: object
          properties:
            flight_id:
              type: string
              format: uuid
            sequence_order:
              type: integer
              example: 1
        - $ref: '#/components/schemas/FlightInput'
...
    HotelDetail:
      allOf:
        - type: object
          properties:
            hotel_id:
              type: string
              format: uuid
            sequence_order:
              type: integer
              example: 1
        - $ref: '#/components/schemas/HotelInput'
...
    ActivityDetail:
      allOf:
        - type: object
          properties:
            activity_id:
              type: string
              format: uuid
            sequence_order:
              type: integer
              example: 1
        - $ref: '#/components/schemas/ActivityInput'

    # ── Approvals ─────────────────────────────
    ApprovalRecord:
      type: object
      properties:
        approval_id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approved, rejected]
          example: "rejected"
        rejection_reason:
          type: string
          nullable: true
          example: "Price does not reflect the stated inclusions. Please itemise all costs clearly."
        reviewed_at:
          type: string
          format: date-time
          example: "2025-03-10T14:22:00+10:00"
...
    MediaItem:
      type: object
      properties:
        media_id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        media_type:
          type: string
          enum: [image, video]
          example: "image"
        url:
          type: string
          format: uri
          example: "https://storage.supabase.co/object/public/media/pkg-00001/bali-rice-terrace.jpg"
        thumbnail_url:
          type: string
          format: uri
          nullable: true
          example: "https://storage.supabase.co/object/public/media/pkg-00001/bali-rice-terrace-thumb.jpg"
        caption:
          type: string
          nullable: true
          example: "Rice terraces at sunrise"
        is_cover:
          type: boolean
          description: Whether this item is the cover/thumbnail for the package.
          example: true
        sort_order:
          type: integer
          example: 1
        uploaded_at:
          type: string
          format: date-time

    # ── Detail-page extras ─────────────────────
    PackageDay:
      type: object
      properties:
        id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        day_number:
          type: integer
          minimum: 1
          example: 1
        title:
          type: string
          nullable: true
          example: "Arrive in Tokyo"
        summary:
          type: string
          nullable: true
          example: "Land at Narita, check in near Shinjuku and ease in with an evening izakaya crawl."
    PackageCreator:
      type: object
      description: Creator profile embedded in marketplace detail responses.
      properties:
        full_name:
          type: string
          example: "Mia Tanaka"
        avatar_url:
          type: string
          format: uri
          nullable: true
        influencer_profiles:
          type: array
          items:
            type: object
            properties:
              bio:
                type: string
                nullable: true
              instagram_handle:
                type: string
                nullable: true
                example: "@mia.eats.world"
              tiktok_handle:
                type: string
                nullable: true
                example: "@mia.eats"
              follower_count:
                type: integer
                example: 482000
              verified:
                type: boolean
    MediaUploadResponse:
      type: object
      properties:
        media_id:
          type: string
          format: uuid
        package_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        media_type:
          type: string
          enum: [image, video]
        filename:
          type: string
        file_size_bytes:
          type: integer
        uploaded_at:
          type: string
          format: date-time

    # ── AI ────────────────────────────────────
...
    MarketplacePackageSummary:
      type: object
      description: Condensed package view for marketplace listing cards.
      properties:
        package_id:
          type: string
          format: uuid
        title:
          type: string
          example: "10-Day Bali Escape: Culture, Surf & Cuisine"
        destination_country:
          type: string
          nullable: true
          example: "Indonesia"
        destination_city:
          type: string
          nullable: true
          example: "Bali"
        duration_days:
          type: integer
          nullable: true
          example: 10
        base_price_aud:
          type: integer
          example: 3500
        cover_image_url:
          type: string
          format: uri
          nullable: true
        tags:
          type: array
          items:
            type: string
        influencer:
          type: object
          properties:
            display_name:
              type: string
              example: "Jessica Travels"
            instagram_handle:
              type: string
              nullable: true
              example: "@jessicatravelsau"
            follower_count:
              type: integer
              nullable: true
              example: 85000
        published_at:
          type: string
          format: date-time

    # ── Pagination ────────────────────────────
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items matching the query.
          example: 142
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8

    # ── Errors ────────────────────────────────
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: Machine-readable error identifier.
          example: "PACKAGE_NOT_FOUND"
        message:
          type: string
          description: Human-readable error description (English, Australia).
          example: "The requested travel package could not be found."
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: Additional context (e.g. validation field errors).
          example:
            field: "base_price_aud"
            issue: "Must be a positive integer (whole AUD dollars)."

  # ─────────────────────────────────────────────
  #  REUSABLE PARAMETERS
  # ─────────────────────────────────────────────
  parameters:
    PackageIdPath:
      name: package_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the travel package.
      example: "pkg-00001-uuid-example-here"
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for paginated results.
    PerPageQuery:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page.

  # ─────────────────────────────────────────────
  #  REUSABLE RESPONSES
  # ─────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "UNAUTHORIZED"
            message: "A valid Bearer token is required to access this resource."
    Forbidden:
      description: Authenticated user does not have the required role for this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "FORBIDDEN"
            message: "Your account role does not permit this action."
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "NOT_FOUND"
            message: "Resource not found."
...
paths:
...
  /media/{package_id}:
    parameters:
      - $ref: '#/components/parameters/PackageIdPath'
    get:
      tags: [media]
      summary: List all media attached to a package
      description: |
        Returns all media items (images and videos) for a given package,
        ordered by `sort_order` ascending.

        **Roles:** `influencer` (own packages), `admin`
        **FR:** FR-I-104
      operationId: listPackageMedia
      responses:
        "200":
          description: Ordered list of media items.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaItem'
                  total:
                    type: integer
                    example: 7
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /media/upload:
    post:
      tags: [media]
      summary: Upload a media file to a package
      description: |
        Uploads an image or video file to Supabase Storage and associates it
        with the specified package. The package must be in `draft` or `rejected`
        status.

        **Supported formats:**
        - Images: `image/jpeg`, `image/png`, `image/webp` (max 10 MB)
        - Videos: `video/mp4`, `video/quicktime` (max 200 MB)

        Set `is_cover=true` to designate this as the cover image/thumbnail.
        Any previous cover is automatically demoted.

        **Roles:** `influencer` (own packages only)
        **FR:** FR-I-104
      operationId: uploadMedia
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [package_id, file]
              properties:
                package_id:
                  type: string
                  format: uuid
                  description: ID of the package to attach this media to.
                file:
                  type: string
                  format: binary
                  description: The media file to upload.
                is_cover:
                  type: boolean
                  default: false
                  description: Set to true to use this as the package cover image.
                sort_order:
                  type: integer
                  default: 0
                  description: Display order position (lower = earlier).
      responses:
        "201":
          description: Media uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadResponse'
              example:
                media_id: "med-uuid-example-here"
                package_id: "pkg-00001-uuid-example-here"
                url: "https://storage.supabase.co/object/public/media/pkg-00001/bali-rice-terrace.jpg"
                media_type: "image"
                filename: "bali-rice-terrace.jpg"
                file_size_bytes: 2457600
                uploaded_at: "2025-03-01T10:15:00+10:00"
        "400":
          description: Unsupported file type or file exceeds size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "UNSUPPORTED_FILE_TYPE"
                message: "Only JPEG, PNG, WebP images and MP4/MOV videos are accepted."
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Package is not in an editable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "PACKAGE_NOT_EDITABLE"
                message: "Media can only be uploaded to packages in 'draft' or 'rejected' status."
  /media/{media_id}:
    parameters:
      - name: media_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Unique identifier of the media item.
    delete:
      tags: [media]
      summary: Delete a media item
      description: |
        Removes a media item from Supabase Storage and the database.
        The associated package must be in `draft` or `rejected` status.

        **Roles:** `influencer` (own packages only)
        **FR:** FR-I-104
      operationId: deleteMedia
      responses:
        "204":
          description: Media deleted successfully. No response body.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Package is not in an editable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "PACKAGE_NOT_EDITABLE"
                message: "Media can only be deleted from packages in 'draft' or 'rejected' status."

  # ══════════════════════════════════════════
  #  AI
  # ══════════════════════════════════════════
...
  /marketplace/packages:
    get:
      tags: [marketplace]
      summary: Browse live packages in the marketplace
      description: |
        Returns a paginated list of all packages with status `live`, visible to
        any visitor — authentication is not required.

        Supports filtering by destination, price range, duration, and tags.
        Default sort is `published_at` descending (newest listings first).

        **Roles:** Public (no authentication required), `customer`
        **FR:** FR-C-100, FR-C-101
      operationId: listMarketplacePackages
      security: []
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - name: destination_country
          in: query
          schema:
            type: string
          description: Filter by destination country name (case-insensitive, partial match).
          example: "Indonesia"
        - name: destination_city
          in: query
          schema:
            type: string
          description: Filter by destination city (case-insensitive, partial match).
          example: "Bali"
        - name: min_price_aud
          in: query
          schema:
            type: integer
            minimum: 0
          description: Minimum price filter in whole AUD dollars (inclusive).
          example: 100000
        - name: max_price_aud
          in: query
          schema:
            type: integer
            minimum: 0
          description: Maximum price filter in whole AUD dollars (inclusive).
          example: 500000
        - name: min_nights
          in: query
          schema:
            type: integer
            minimum: 1
          description: Minimum trip duration in nights.
          example: 7
        - name: max_nights
          in: query
          schema:
            type: integer
            minimum: 1
          description: Maximum trip duration in nights.
          example: 14
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Filter by one or more tags (comma-separated). Returns packages matching ANY tag.
          example: "beach,surf"
        - name: sort
          in: query
          schema:
            type: string
            enum: [published_at_desc, price_asc, price_desc, duration_asc]
            default: published_at_desc
      responses:
        "200":
          description: Paginated list of live marketplace packages.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MarketplacePackageSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - package_id: "pkg-00001-uuid-example-here"
                    title: "10-Day Bali Escape: Culture, Surf & Cuisine"
                    destination_country: "Indonesia"
                    destination_city: "Bali"
                    duration_days: 10
                    base_price_aud: 3500
                    cover_image_url: "https://storage.supabase.co/object/public/media/pkg-00001/cover.jpg"
                    tags: ["beach", "culture", "food", "surf"]
                    influencer:
                      display_name: "Jessica Travels"
                      instagram_handle: "@jessicatravelsau"
                      follower_count: 85000
                    published_at: "2025-03-12T09:00:00+10:00"
                meta:
                  total: 42
                  page: 1
                  per_page: 20
                  total_pages: 3
  /marketplace/packages/{package_id}:
    parameters:
      - $ref: '#/components/parameters/PackageIdPath'
    get:
      tags: [marketplace]
      summary: Get full details of a live marketplace package
      description: |
        Returns the complete public view of a package, including all itinerary
        components (flights, hotels, activities) and media. Only packages with
        status `live` are accessible via this endpoint.

        **Roles:** Public (no authentication required), `customer`
        **FR:** FR-C-100
      operationId: getMarketplacePackage
      security: []
      responses:
        "200":
          description: Full package detail for marketplace display.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TravelPackageDetail'
        "404":
          description: Package not found or not in `live` status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "PACKAGE_NOT_FOUND"
                message: "The requested travel package could not be found or is not currently listed."
  /marketplace/search:
    get:
      tags: [marketplace]
      summary: Full-text search across live packages
      description: |
        Performs a full-text search against package `title`, `description`,
        `destination_city`, and `tags` fields. Returns only `live` packages.

        Results are ranked by text relevance score, with secondary sort by
        `published_at` descending.

        This endpoint supports the same filter parameters as `GET /marketplace/packages`
        and may be combined with the `q` query parameter.

        **Roles:** Public (no authentication required), `customer`
        **FR:** FR-C-101, FR-C-102
      operationId: searchMarketplace
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 200
          description: Full-text search query.
          example: "Bali surfing beginners"
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - name: destination_country
          in: query
          schema:
            type: string
          example: "Indonesia"
        - name: min_price_aud
          in: query
          schema:
            type: integer
            minimum: 0
          example: 100000
        - name: max_price_aud
          in: query
          schema:
            type: integer
            minimum: 0
          example: 500000
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          example: "surf,beach"
      responses:
        "200":
          description: Search results, ranked by relevance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                    description: The search query as received.
                    example: "Bali surfing beginners"
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/MarketplacePackageSummary'
                        - type: object
                          properties:
                            relevance_score:
                              type: number
                              format: float
                              description: Text search relevance score (higher = more relevant).
                              example: 0.847
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        "400":
          description: Query string too short or malformed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: "SEARCH_QUERY_TOO_SHORT"
                message: "The search query must be at least 2 characters long."